light.transition now supports float instead of int in order to be able to perform faster transitions (#6163)

+++ homeassistant/components/light/hue.py
@@ -376,7 +376,7 @@ class HueLight(Light):
         command = {'on': True}
 
         if ATTR_TRANSITION in kwargs:
-            command['transitiontime'] = kwargs[ATTR_TRANSITION] * 10
+            command['transitiontime'] = int(kwargs[ATTR_TRANSITION] * 10)
 
         if ATTR_XY_COLOR in kwargs:
             command['xy'] = kwargs[ATTR_XY_COLOR]
@@ -422,7 +422,10 @@ class HueLight(Light):
         if ATTR_TRANSITION in kwargs:
             # Transition time is in 1/10th seconds and cannot exceed
             # 900 seconds.
-            command['transitiontime'] = min(9000, kwargs[ATTR_TRANSITION] * 10)
+            command['transitiontime'] = min(
+                9000,
+                int(kwargs[ATTR_TRANSITION] * 10)
+            )
 
         flash = kwargs.get(ATTR_FLASH)
